Refactor some shape widget code in new Instrument View (#41196)
